Cambiando valor al guardar

+++ app/src/main/java/com/call/prestamoamigo/view/Prestamo/PrestamoViewModel.kt
@@ -36,7 +36,7 @@ class PrestamoViewModel @Inject constructor(
                     fecha = fecha,
                     personaId = personaIdentification.toInt(),
                     concepto = concepto,
-                    monto = monto.toString().toFloat(),
+                    monto = monto.toString().toDouble(),
                     vence = vence,
                     balance = monto.toString().toDouble(),
                  )
